Fix first Linux/CMake-4 CI failures: dependent template keyword, suffixed math, policy minimum
- Script.hpp: script->template ExecuteStateTracker<T>() (GCC/Clang require the
  keyword on a dependent object; MSVC accepted the omission). Drop the
  self-friend declaration GCC warns about.
- std::sqrtf/floorf/roundf/remainderf -> unsuffixed std overloads; libstdc++ 13
  does not declare the f-suffixed names in std.
- printf of an enum class in BitfsOscFinal cast to int (-Wformat).
- CMAKE_POLICY_VERSION_MINIMUM 3.5 so CMake 4.x accepts nlohmann/json 3.11.2.
- CI: explicit bash shell on Linux steps so pipefail stops tee from hiding a
  failed build. All catalogued in docs/compilers.md.

+++ tasfw-scripts/src/bitfs-pyramid-osc/BitFsPyramidOscillation_GetMinimumDownhillWalkingAngle.cpp
@@ -40,7 +40,9 @@ bool BitFsPyramidOscillation_GetMinimumDownhillWalkingAngle::execution()
 	auto floor = &pyramid->surfaces[1][_floorId];
 	short floorAngle = atan2s(floor->normal.z, floor->normal.x);
 	CustomStatus.isSlope = PyramidUpdateMem::FloorIsSlope(floor, marioState->action);
-	CustomStatus.steepness = std::sqrtf(floor->normal.x * floor->normal.x + floor->normal.z * floor->normal.z);
+	// std::sqrt(float) overload, not std::sqrtf: libstdc++ 13 does not declare the f-suffixed
+	// names in namespace std (docs/compilers.md). Same instruction, same result.
+	CustomStatus.steepness = std::sqrt(floor->normal.x * floor->normal.x + floor->normal.z * floor->normal.z);
 
 	// m->floorAngle - m->faceAngle[1] >= -0x3FFF && m->floorAngle -
 	// m->faceAngle[1] <= 0x3FFF
